# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: SAMT Unified Backend API
  version: 1.0.0
  description: Generated by scripts/generate-openapi.mjs from service OpenAPI docs exposed via API Gateway.
servers:
  - url: http://localhost:9080
...
paths:
...
  /api/admin/users/{userId}:
    delete:
      tags:
        - Admin
      summary: Soft delete user
      description: Soft delete a user (set deleted_at, revoke tokens). Does not permanently remove data.
      operationId: deleteUser
      parameters:
        - name: userId
          in: path
          description: User ID
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
      responses:
        '200':
          description: User deleted
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/identity_ApiResponseAdminActionResponse'
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '405':
          $ref: '#/components/responses/Error405'
        '409':
          $ref: '#/components/responses/Error409'
        '415':
          $ref: '#/components/responses/Error415'
        '500':
          $ref: '#/components/responses/Error500'
      security:
        - bearerAuth: []
  /api/users/{userId}:
    get:
      tags:
        - user-controller
      operationId: getUserById
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/userGroup_ApiEnvelope_addMember'
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/userGroup_ApiEnvelope_getUserById'
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '405':
          $ref: '#/components/responses/Error405'
        '409':
          $ref: '#/components/responses/Error409'
        '415':
          $ref: '#/components/responses/Error415'
        '500':
          $ref: '#/components/responses/Error500'
    put:
      tags:
        - user-controller
      operationId: updateUser
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/userGroup_UpdateUserRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/userGroup_ApiEnvelope_updateUser'
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '405':
          $ref: '#/components/responses/Error405'
        '409':
          $ref: '#/components/responses/Error409'
        '415':
          $ref: '#/components/responses/Error415'
        '500':
          $ref: '#/components/responses/Error500'
...
  /api/groups/{groupId}:
    get:
      tags:
        - group-controller
      operationId: getGroupById
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/userGroup_ApiEnvelope_getGroupById'
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '405':
          $ref: '#/components/responses/Error405'
        '409':
          $ref: '#/components/responses/Error409'
        '415':
          $ref: '#/components/responses/Error415'
        '500':
          $ref: '#/components/responses/Error500'
    put:
      tags:
        - group-controller
      operationId: updateGroup
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/userGroup_UpdateGroupRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/userGroup_ApiEnvelope_updateGroup'
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '405':
          $ref: '#/components/responses/Error405'
        '409':
          $ref: '#/components/responses/Error409'
        '415':
          $ref: '#/components/responses/Error415'
        '500':
          $ref: '#/components/responses/Error500'
    delete:
      tags:
        - group-controller
      operationId: deleteGroup
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
      responses:
        '200':
          description: OK
        '204':
          description: No Content
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '405':
          $ref: '#/components/responses/Error405'
        '409':
          $ref: '#/components/responses/Error409'
        '415':
          $ref: '#/components/responses/Error415'
        '500':
          $ref: '#/components/responses/Error500'
...
  /api/groups:
    get:
      tags:
        - group-controller
      operationId: listGroups
      parameters:
        - name: page
          in: query
          required: false
          schema:
            anyOf:
              - type: integer
                format: int32
                default: 0
                minimum: 0
              - type: string
                enum:
                  - ''
            default: 0
          allowEmptyValue: true
        - name: size
          in: query
          required: false
          schema:
            anyOf:
              - type: integer
                format: int32
                default: 20
                maximum: 100
                minimum: 1
              - type: string
                enum:
                  - ''
            default: 20
          allowEmptyValue: true
        - name: semesterId
          in: query
          required: false
          schema:
            type: integer
            format: int64
            minimum: 1
        - name: lecturerId
          in: query
          required: false
          schema:
            type: integer
            format: int64
            minimum: 1
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/userGroup_ApiEnvelope_listGroups'
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '405':
          $ref: '#/components/responses/Error405'
        '409':
          $ref: '#/components/responses/Error409'
        '415':
          $ref: '#/components/responses/Error415'
        '500':
          $ref: '#/components/responses/Error500'
    post:
      tags:
        - group-controller
      operationId: createGroup
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/userGroup_CreateGroupRequest'
        required: true
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/userGroup_ApiEnvelope_createGroup'
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '405':
          $ref: '#/components/responses/Error405'
        '409':
          $ref: '#/components/responses/Error409'
        '415':
          $ref: '#/components/responses/Error415'
        '500':
          $ref: '#/components/responses/Error500'
...
  /api/groups/{groupId}/lecturer:
    patch:
      tags:
        - group-controller
      operationId: updateGroupLecturer
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/userGroup_UpdateLecturerRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/userGroup_ApiEnvelope_updateGroupLecturer'
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '405':
          $ref: '#/components/responses/Error405'
        '409':
          $ref: '#/components/responses/Error409'
        '415':
          $ref: '#/components/responses/Error415'
        '500':
          $ref: '#/components/responses/Error500'
  /api/users:
    get:
      tags:
        - user-controller
      operationId: listUsers
      parameters:
        - name: page
          in: query
          required: false
          schema:
            anyOf:
              - type: integer
                format: int32
                default: 0
                minimum: 0
              - type: string
                enum:
                  - ''
            default: 0
          allowEmptyValue: true
        - name: size
          in: query
          required: false
          schema:
            anyOf:
              - type: integer
                format: int32
                default: 20
                maximum: 100
                minimum: 1
              - type: string
                enum:
                  - ''
            default: 20
          allowEmptyValue: true
        - name: status
          in: query
          required: false
          schema:
            type: string
        - name: role
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/userGroup_ApiEnvelope_listUsers'
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '405':
          $ref: '#/components/responses/Error405'
        '409':
          $ref: '#/components/responses/Error409'
        '415':
          $ref: '#/components/responses/Error415'
        '500':
          $ref: '#/components/responses/Error500'
  /api/users/{userId}/groups:
    get:
      tags:
        - user-controller
      operationId: getUserGroups
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
        - name: semester
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/userGroup_ApiEnvelope_getUserGroups'
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '405':
          $ref: '#/components/responses/Error405'
        '409':
          $ref: '#/components/responses/Error409'
        '415':
          $ref: '#/components/responses/Error415'
        '500':
          $ref: '#/components/responses/Error500'
...
  /api/project-configs/group/{groupId}:
    get:
      tags:
        - project-config-controller
      operationId: getConfigByGroupId
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/projectConfig_ConfigEnvelope'
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '405':
          $ref: '#/components/responses/Error405'
        '409':
          $ref: '#/components/responses/Error409'
        '415':
          $ref: '#/components/responses/Error415'
        '500':
          $ref: '#/components/responses/Error500'
...
components:
  schemas:
...
    identity_AdminActionResponse:
      type: object
      properties:
        message:
          type: string
        userId:
          type: string
    identity_ApiResponseAdminActionResponse:
      type: object
      properties:
        timestamp:
          type: string
        status:
          type: integer
          format: int32
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/identity_AdminActionResponse'
        error:
          type: string
        message:
          type: string
        path:
          type: string
        correlationId:
          type: string
        degraded:
          type: boolean
...
    userGroup_UpdateUserRequest:
      type: object
      properties:
        fullName:
          type: string
          maxLength: 100
          minLength: 2
          pattern: '^[a-zA-ZÀ-ỹ]+(?: [a-zA-ZÀ-ỹ]+)*$'
      required:
        - fullName
    userGroup_UserResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        email:
          type: string
        fullName:
          type: string
        status:
          type: string
        roles:
          type: array
          items:
            type: string
...
    userGroup_UpdateGroupRequest:
      type: object
      properties:
        groupName:
          type: string
          maxLength: 50
          minLength: 3
          pattern: ^[A-Z]{2,4}[0-9]{2,4}-G[0-9]+$
        lecturerId:
          type: integer
          format: int64
          minimum: 1
      required:
        - groupName
        - lecturerId
...
    userGroup_MemberResponse:
      type: object
      properties:
        userId:
          type: integer
          format: int64
        groupId:
          type: integer
          format: int64
        semesterId:
          type: integer
          format: int64
        groupRole:
          type: string
          enum:
            - LEADER
            - MEMBER
        joinedAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    userGroup_CreateGroupRequest:
      type: object
      properties:
        groupName:
          type: string
          maxLength: 50
          minLength: 3
          pattern: ^[A-Z]{2,4}[0-9]{2,4}-G[0-9]+$
        semesterId:
          type: integer
          format: int64
          minimum: 1
        lecturerId:
          type: integer
          format: int64
          minimum: 1
      required:
        - groupName
        - semesterId
        - lecturerId
...
    userGroup_UpdateLecturerRequest:
      type: object
      properties:
        lecturerId:
          type: integer
          format: int64
          minimum: 1
      required:
        - lecturerId
...
    userGroup_UserGroupsResponse:
      type: object
      properties:
        userId:
          type: integer
          format: int64
        groups:
          type: array
          items:
            $ref: '#/components/schemas/userGroup_GroupInfo'
...
    userGroup_GroupDetailResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        groupName:
          type: string
        semesterId:
          type: integer
          format: int64
        semesterCode:
          type: string
        lecturer:
          $ref: '#/components/schemas/userGroup_LecturerInfo'
        members:
          type: array
          items:
            $ref: '#/components/schemas/userGroup_MemberInfo'
        memberCount:
          type: integer
          format: int32
...
    StandardError:
      type: object
      properties:
        type:
          type: string
          example: https://api.example.com/errors/validation-error
        title:
          type: string
          example: Validation failed
        status:
          type: integer
          format: int32
          example: 400
        detail:
          type: string
          example: Invalid UUID
        instance:
          type: string
          example: /api/users/not-a-number
        timestamp:
          type: string
          format: date-time
...
    projectConfig_ConfigEnvelope:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/projectConfig_ConfigData'
        timestamp:
          type: string
          format: date-time
...
    userGroup_ApiEnvelope_getUserById:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
          format: int32
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/userGroup_UserResponse'
        error:
          type: string
        message:
          type: string
        path:
          type: string
        correlationId:
          type: string
        degraded:
          type: boolean
    userGroup_ApiEnvelope_updateUser:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
          format: int32
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/userGroup_UserResponse'
        error:
          type: string
        message:
          type: string
        path:
          type: string
        correlationId:
          type: string
        degraded:
          type: boolean
...
    userGroup_ApiEnvelope_getGroupById:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
          format: int32
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/userGroup_GroupDetailResponse'
        error:
          type: string
        message:
          type: string
        path:
          type: string
        correlationId:
          type: string
        degraded:
          type: boolean
    userGroup_ApiEnvelope_updateGroup:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
          format: int32
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/userGroup_GroupResponse'
        error:
          type: string
        message:
          type: string
        path:
          type: string
        correlationId:
          type: string
        degraded:
          type: boolean
...
    userGroup_ApiEnvelope_listGroups:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
          format: int32
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/userGroup_PageResponseGroupListResponse'
        error:
          type: string
        message:
          type: string
        path:
          type: string
        correlationId:
          type: string
        degraded:
          type: boolean
    userGroup_ApiEnvelope_createGroup:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
          format: int32
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/userGroup_GroupResponse'
        error:
          type: string
        message:
          type: string
        path:
          type: string
        correlationId:
          type: string
        degraded:
          type: boolean
...
    userGroup_ApiEnvelope_addMember:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
          format: int32
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/userGroup_MemberResponse'
        error:
          type: string
        message:
          type: string
        path:
          type: string
        correlationId:
          type: string
        degraded:
          type: boolean
    userGroup_ApiEnvelope_updateGroupLecturer:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
          format: int32
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/userGroup_GroupResponse'
        error:
          type: string
        message:
          type: string
        path:
          type: string
        correlationId:
          type: string
        degraded:
          type: boolean
    userGroup_ApiEnvelope_listUsers:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
          format: int32
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/userGroup_PageResponseUserResponse'
        error:
          type: string
        message:
          type: string
        path:
          type: string
        correlationId:
          type: string
        degraded:
          type: boolean
    userGroup_ApiEnvelope_getUserGroups:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
          format: int32
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/userGroup_UserGroupsResponse'
        error:
          type: string
        message:
          type: string
        path:
          type: string
        correlationId:
          type: string
        degraded:
          type: boolean
...
  responses:
    Error400:
      description: 400 error
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/StandardError'
    Error401:
      description: 401 error
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/StandardError'
    Error403:
      description: 403 error
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/StandardError'
    Error404:
      description: 404 error
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/StandardError'
    Error405:
      description: 405 error
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/StandardError'
    Error409:
      description: 409 error
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/StandardError'
    Error415:
      description: 415 error
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/StandardError'
    Error500:
      description: 500 error
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/StandardError'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    identity_bearerAuth:
      type: http
      description: Enter JWT access token
      scheme: bearer
      bearerFormat: JWT
    userGroup_bearerAuth:
      type: http
      description: JWT token from identity-service
      name: bearerAuth
      scheme: bearer
      bearerFormat: JWT
    projectConfig_Bearer Authentication:
      type: http
      description: Enter your JWT token from Identity Service
      scheme: bearer
      bearerFormat: JWT
    sync_bearerAuth:
      type: http
      description: Paste the JWT only. Swagger UI adds the Bearer prefix automatically.
      name: Authorization
      in: header
      scheme: bearer
      bearerFormat: JWT
    report_bearerAuth:
      type: http
      description: Paste the JWT only. Swagger UI adds the Bearer prefix automatically.
      name: Authorization
      in: header
      scheme: bearer
      bearerFormat: JWT
...